use https://github.com/home-assistant/example-custom-config/blob/master/custom_components/example_sensor/sensor.py instead of blueprint
